Implement the project desing in Calculator.js file
